+++ src/hybrid/hybrid_request.c
@@ -185,7 +185,7 @@ void HybridRequest_InitArgsCursor(HybridRequest *req, ArgsCursor *ac, RedisModul
   req->args = rm_calloc(argc, sizeof(*req->args));
   req->nargs = argc;
   // Copy the arguments into an owned array of sds strings
-  for (size_t ii = 0; ii < argc; ++ii) {
+  for (int ii = 0; ii < argc; ++ii) {
     size_t n;
     const char *s = RedisModule_StringPtrLen(argv[ii], &n);
     req->args[ii] = sdsnewlen(s, n);
