+++ src/index_result/index_result.c
@@ -129,8 +129,8 @@ int IndexResult_MinOffsetDelta(const RSIndexResult *r) {
     uint32_t p1 = v1.Next(v1.ctx, NULL);
     uint32_t p2 = v2.Next(v2.ctx, NULL);
     int cd = __absdelta(p2, p1);
-    while (cd > 1 && p1 != RS_OFFSETVECTOR_EOF && p2 != RS_OFFSETVECTOR_EOF) {
-      cd = MIN(__absdelta(p2, p1), cd);
+    while ((uint32_t)cd > 1 && p1 != RS_OFFSETVECTOR_EOF && p2 != RS_OFFSETVECTOR_EOF) {
+      cd = MIN(__absdelta(p2, p1), (uint32_t)cd);
       if (p2 > p1) {
         p1 = v1.Next(v1.ctx, NULL);
       } else {
@@ -159,7 +159,7 @@ void result_GetMatchedTerms(const RSIndexResult *r, RSQueryTerm *arr[], size_t c
       // and skip the tag check on the next line.
       const RSAggregateResult *agg = IndexResult_AggregateRefUnchecked(r);
       AggregateRecordsSlice children = AggregateResult_GetRecordsSlice(agg);
-      for (int i = 0; i < children.len; i++) {
+      for (size_t i = 0; i < children.len; i++) {
         result_GetMatchedTerms(children.ptr[i], arr, cap, len);
       }
